Fix `show-truffle-env`
Prior to this change, the `show-truffle-env` command would fail under zsh
on macos when running truffle version. The underlying cause of this
failure is that aliases aren't expanded in noninteractive sessions (such
as when sourcing a script). This means that the `truffle version`
command in `show-truffle-env` was likely always running the globally
installed `truffle` when displaying its version. I didn't have a
globally installed `truffle`, so in my case I was seeing an error.

This change switches away from using aliases for the `truffle` and
`db-kit` commands, and uses functions. With this change, scripts
executed in noninteractive subshells will still inherit the configurable
`truffle` command provided by this dotfile.